Combine button components
