--- src/client/toggle-button.module.css
@@ -1,31 +1,0 @@
-.button {
-  background: none;
-  border: none;
-  display: flex;
-  align-items: center;
-  border: 1px solid transparent;
-  opacity: 0.7;
-  outline: 0 !important;
-  margin-left: 3px;
-  margin-right: 3px;
-  padding: 3px;
-  cursor: pointer;
-}
-
-.button + .button {
-  margin-left: 0;
-}
-
-.button:hover {
-  opacity: 1;
-}
-
-.button[aria-checked='true'] {
-  background: var(--vscode-inputOption-activeBackground);
-  border: 1px solid var(--vscode-inputOption-activeBorder);
-  opacity: 1;
-}
-
-.button > svg {
-  width: 1em;
-}
